Trying to fix Patrick's problems when launching getdp on Windows...
When gmsh is started from the command line on Windows, cygwin does
not (for some obscure reason) prepend '/cygdrive/x/' to the path. Since
we check for '/cygdrive/' in SystemCall() in order to sanitize the
path before passing it to CreateProcess(), the system call fails.

I think the right way to deal with this is to actually remove the
path change in SystemCall() (since we only change the leading part),
and use the official cygwin API to change the path of every piece
that will eventually end up in the argument of CreateProcess().
And that's what "FixWindowsPath()" now does...


git-svn-id: https://geuz.org/svn/gmsh/trunk@2840 59c8cabf-10d6-4a28-b43f-3936b85956c4