bug fix from peter.wainwright@ieee.org
Amazingly, I have discovered a potentially serious bug in the routines
which perform edge collapse in the BDS_Mesh.

Owing to this, it is possible for a point to collapse onto a linear
boundary creating slivers (triangles with collinear points).

The problem is that the sides of the triangle, a and b, are not
recalculated after the point is moved.  Thus, area_final and area_init
are always equal!  So the test for slivers is defeated.


git-svn-id: https://geuz.org/svn/gmsh/trunk@5987 59c8cabf-10d6-4a28-b43f-3936b85956c4