Remove ElasticFields. Now the solver has a vector<partDomain*> and not vector<partDomain> (I use Domain in place of Field because field is for the computation of unkwowns). A component of this vector can be a ElasticField but also an other type of Domain. The Linear and Bilinear terms are included in the domain which allow to use properly the iterative scheme with different terms.
The Gauss Integration Rule is also included in the Domain which allow to
choose this integration. In case of a dgLinearShellDomain
underintegration is chosen for second order shape functions and full
integration is performed for third order shape functions. NB: A lua
intereaction can be added to allow an user's choice of the Integration
Rule.

Update Benchmarks
 


git-svn-id: https://geuz.org/svn/gmsh/trunk@7973 59c8cabf-10d6-4a28-b43f-3936b85956c4