Allow to not compile the Gmsh executable (very usefull when using gmsh as a library)
Add a default implementation of fullMatrix<double>::invert and
fullMatrix<double>::determinant whn Lapack is not available. The algorirthms are bad but they allow to use polynomialBasis without installing BLAS and Lapack. The inverse is not computed a lot a time and the matrix is often not big in this case so there is no need of better algorithms


git-svn-id: https://geuz.org/svn/gmsh/trunk@15222 59c8cabf-10d6-4a28-b43f-3936b85956c4