Search functions in OctreePost can now be given the node coordinates of an element.
In that case we search for all the matches in the octree, and we try to return amongst those the element that matches the input coordinates.

We use this in MathEval so we can reliably make operations on discontinuous datasets (e.g. edge element vector fields), on identical meshes that were possibly reordered.



git-svn-id: https://geuz.org/svn/gmsh/trunk@17382 59c8cabf-10d6-4a28-b43f-3936b85956c4